# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5f46db3167541c707d179879
  title: appvaa
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/teachers/{id}":
    delete:
      x-skaffolder-id: 5f46eac167541c707d17997e
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5f46e3fa67541c707d179914
      x-skaffolder-resource: Teachers
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Teachers
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5f46eac167541c707d179973
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5f46e3fa67541c707d179914
      x-skaffolder-resource: Teachers
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Teachers
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teachers"
    post:
      x-skaffolder-id: 5f46eac167541c707d179974
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5f46e3fa67541c707d179914
      x-skaffolder-resource: Teachers
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Teachers
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teachers"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Teachers"
  "/teachers/findBy_class/{key}":
    get:
      x-skaffolder-id: 5f46eb1867541c707d1799cf
      x-skaffolder-name: findBy_class
      x-skaffolder-id-resource: 5f46e3fa67541c707d179914
      x-skaffolder-resource: Teachers
      x-skaffolder-crudAction: findBy_class
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_class
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_class/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_class
      tags:
        - Teachers
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teachers"
  "/teachers/findBy_club/{key}":
    get:
      x-skaffolder-id: 5f46ead567541c707d179998
      x-skaffolder-name: findBy_club
      x-skaffolder-id-resource: 5f46e3fa67541c707d179914
      x-skaffolder-resource: Teachers
      x-skaffolder-crudAction: findBy_club
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_club
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_club/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_club
      tags:
        - Teachers
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teachers"
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Teachers:
      x-skaffolder-id: 5f46e3fa67541c707d179914
      x-skaffolder-id-db: 5f46db3167541c707d17987a
      x-skaffolder-id-entity: 5f46e3f967541c707d17990c
      x-skaffolder-url: /teachers
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        auth_classes_students:
          type: boolean
          x-skaffolder-id-attr: 5f46e3f967541c707d17990d
          x-skaffolder-type: Boolean
          x-skaffolder-required: true
        date_of_birth:
          type: integer
          x-skaffolder-id-attr: 5f46e3f967541c707d17990e
          x-skaffolder-type: Date
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 5f46e3f967541c707d17990f
          x-skaffolder-type: String
          x-skaffolder-required: true
        partner:
          type: boolean
          x-skaffolder-id-attr: 5f46e3f967541c707d179910
          x-skaffolder-type: Boolean
          x-skaffolder-required: true
        password:
          type: string
          x-skaffolder-id-attr: 5f46e3f967541c707d179911
          x-skaffolder-type: String
          x-skaffolder-required: true
        surname:
          type: string
          x-skaffolder-id-attr: 5f46e3f967541c707d179912
          x-skaffolder-type: String
          x-skaffolder-required: true
        username:
          type: string
          x-skaffolder-id-attr: 5f46e3f967541c707d179913
          x-skaffolder-type: String
          x-skaffolder-required: true
          x-skaffolder-unique: true
      x-skaffolder-relations:
        _class:
          x-skaffolder-id: 5f46e62b67541c707d17995d
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5f46e3f967541c707d17990c
          x-skaffolder-ent2: 5f46e59767541c707d179951
        _club:
          x-skaffolder-id: 5f46e62b67541c707d17995e
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5f46e3f967541c707d17990c
          x-skaffolder-ent2: 5f46e51a67541c707d17992b
      required:
        - _id
        - auth_classes_students
        - date_of_birth
        - name
        - partner
        - password
        - surname
        - username
...